chore(tooling): declare the quality toolchain and make the tree green under it
`uv run mypy` was resolving to a user-site mypy on PATH, under a different
interpreter that cannot see this project's dependencies -- because mypy was
never declared as a dependency. It reported ~19 phantom `import-not-found`
errors, and `uv run coverage run -m pytest` failed collection outright with
25 errors. Both signals looked broken, so both were ignored, and `CLAUDE.md`
recorded them as "ambient tools, not yet declared".

Declares mypy, ruff, coverage, pytest-cov and types-defusedxml in
`[dependency-groups] dev`, and declares their rulesets in `pyproject.toml`
rather than inheriting whatever the installed versions default to. Adds the
CI workflow that runs all three on every push, with a 93% coverage floor
(measured: 96%).

Turning the gates on surfaced three real defects, fixed here because
otherwise the very first commit fails the gate it installs:

- `DocumentRendererPort.build` declared `-> Path` while HTML and PDF return
  `Path | None` by design. The port was the side that lied; correcting it
  exposed two callers assuming non-None without saying so -- PdfRendererAdapter
  handed `None` to the soffice adapter, and `stage_build_docx` reported
  ok=True with the literal detail "None".
- `MermaidSvgRenderer` promised "a clean, catchable error" in its docstring
  and leaked a raw `CalledProcessError` whose message is a path dump, with
  mermaid's own diagnostic lost to an uncaptured stderr.
- The composition root swallowed renderer-registration failures with a bare
  `pass`, making a broken renderer indistinguishable from an uninstalled one.

Also writes the README, which was zero bytes, and fills in the placeholder
package description.

mypy: 17 errors -> 0.   ruff: 159 -> 0.   1525 passed, 7 skipped, 96.31%.

+++ src/docs/domain/docx_structure.py
@@ -64,7 +64,7 @@ def non_cover_margin_emu(config: dict[str, Any]) -> dict[str, int]:
     for key in _MARGIN_KEYS:
         value = margins.get(key)
         if isinstance(value, (int, float)):
-            expected[key] = int(round(float(value) * EMU_PER_CM))
+            expected[key] = round(float(value) * EMU_PER_CM)
     return expected
